Check prop for getter before building serializer
Relevant code starts at 2821 - if the value to be serialized is a property, pull the getter method immediately, then do a check - if it's not a property, or it has no getter, will no longer attempt to build a serializer for that property. 

This is to correct a crash caused by the fact that lines 2860 and 2868 (2856 and 2862 in the original file) assumed GetGetMethod will not return as null. If a property has no getter anyways, it doesn't need to be mapped for serialization.